More Bootstrap JQuery
In this exercise we learnt more about Bootstrap's JS component methods
and how they can be used together with JQuery and JavaScript to control
the behavior of a Bootstrap JS component.

We change the buttons (pause and play) that we add before, to a single
one that change the icon and the action when we click it.